refactor: domain으로 옮김
