+++ src/main/java/nextstep/ladder/domain/Player.java
@@ -1,4 +1,4 @@
-package nextstep.ladder;
+package nextstep.ladder.domain;
 
 public class Player {
     private static final String NAME_LENGTH_OVER = "이름은 최대 5글자까지 가능합니다.";
@@ -14,5 +14,9 @@ public class Player {
         if (name.length() > NAME_MAX_LENGTH) {
             throw new IllegalArgumentException(NAME_LENGTH_OVER);
         }
+    }
+
+    public String getName() {
+        return name;
     }
 }
