+++ src/test/java/nextstep/ladder/domain/PlayerTest.java
@@ -1,4 +1,4 @@
-package nextstep.ladder;
+package nextstep.ladder.domain;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
